requirements.txt
# base
PIP >= 21.1.1
# Poetry >= 1.1.4
    # use default installer instead:
    # curl -sSL https://raw.githubusercontent.com/python-poetry/poetry/master/get-poetry.py | python


# EV3
# ---
https://GitHub.com/PyBricks/PyBricks-API/archive/master.zip
# PyBricks >= 3.0.0
# PyBricks-Stubs >= 2.0.2   # use official PyBricks above
Python-EV3Dev2 >= 2.1.0.post1

# EV3Dev Python Tools
EV3DevContext >= 0.36
EV3DevCmd >= 1.0.2
EV3DevLogging >= 1.0.1
EV3DevRPyC >= 1.0.1
EV3Dev2Simulator
    # ev3dev2simulator 2.0.5 depends on pymunk==5.6.0
    # ev3sim 2.1.2 depends on pymunk>=6.0.0
Thonny-EV3Dev >= 1.0.4

# deprecated
# EV3DevLight >= 0.1.5
    # installation error
Python-EV3Dev >= 1.2.0

# Simulators
EV3Sim >= 2.1.7
PyGame >= 2.0.1

# misc
EV3-Python >= 0.0.1
# EV3Dev-C >= 1.0rc2
    # installation error
# LEGO-MicroPython-Extension >= 0.2.1.dev2019110502
    # no available distribution
# LEGO-MicroPython-Skeleton >= 0.1.9.2019103111
    # no available distribution


# NXT
# ---
git+https://GitHub.com/EV3Dev/NXT-Python
# NXT-Python >= 2.2.2
# NXT-Python2 >= 2.2.2
# NXT-Python-Tools >= 1.0.1
    # invalid old syntax

git+http://git.ni.fr.eu.org/pyfantom.git
PyUSB >= 1.1.1

BlockExt-Mindstorms-NXT >= 0.2
# BrickPython >= 0.4
    # installation error
KLego >= 0.98.2.6
JaRaCo.NXT >= 2.1.0

# NXTurtle >= 1.0.4.dev-20110105
    # invalid old syntax


# Non-Mindstorms
# --------------
BrickNil >= 0.9.3
PyLGBst >= 1.2.0
WeDo2 >= 1.1.1
# WeDo >= 1.1.0
    # installation error

# BrickPi
BrickPi-Python >= 1.0

# Networking
AdaFruit-AMPy >= 1.1.0
RPyC
    # ev3devcmd 1.0.2 depends on rpyc==4.1.2

# Linting & Refactoring
PyLint >= 2.8.2
Flake8 >= 3.9.2
MyPy >= 0.812
PyDocStyle >= 6.0.0
PyCodeStyle >= 2.7.0
# Prospector >= 1.3.0
    # *** prospector 1.3.0 depends on pylint==2.5.2 ***
Pylama >= 7.7.1
Bandit >= 1.7.0
Rope >= 0.19.0
